+++ src/components/Navigation.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { useAuth } from "../auth/hooks/useAuth";
+import { useAuth } from "../features/auth/hooks/useAuth";
 
 interface NavigationProps {
   variant?: 'default' | 'dark' | 'download';
